moved files to the correct location
