+++ util/util.py
@@ -45,22 +45,17 @@ def diagnose_network(net, name='network'):
     print(name)
     print(mean)
 
-
-def save_image(image_numpy, image_path, aspect_ratio=1.0):
+#added a flag for end saving
+def save_image(image_numpy, image_path, end = False):
     """Save a numpy image to the disk
 
     Parameters:
         image_numpy (numpy array) -- input numpy array
         image_path (str)          -- the path of the image
     """
-
     image_pil = Image.fromarray(image_numpy)
-    h, w, _ = image_numpy.shape
-
-    if aspect_ratio > 1.0:
-        image_pil = image_pil.resize((h, int(w * aspect_ratio)), Image.BICUBIC)
-    if aspect_ratio < 1.0:
-        image_pil = image_pil.resize((int(h / aspect_ratio), w), Image.BICUBIC)
+    if end:
+        image_pil = image_pil.crop((0, 0, 300, 350)) #((left, top, right, bottom)) 
     image_pil.save(image_path)
