+++ util/visualizer.py
@@ -5,7 +5,8 @@ import ntpath
 import time
 from . import util, html
 from subprocess import Popen, PIPE
-
+#from scipy.misc import imresize
+from skimage.transform import resize
 
 if sys.version_info[0] == 2:
     VisdomExceptionBase = Exception
@@ -36,7 +37,13 @@ def save_images(webpage, visuals, image_path, aspect_ratio=1.0, width=256):
         im = util.tensor2im(im_data)
         image_name = '%s_%s.png' % (name, label)
         save_path = os.path.join(image_dir, image_name)
-        util.save_image(im, save_path, aspect_ratio=aspect_ratio)
+        h, w, _ = im.shape
+        if aspect_ratio > 1.0:
+            im = imresize(im, (h, int(w * aspect_ratio)), interp='bicubic')
+        if aspect_ratio < 1.0:
+            im = imresize(im, (int(h / aspect_ratio), w), interp='bicubic')
+        util.save_image(im, save_path, True)
+
         ims.append(image_name)
         txts.append(label)
         links.append(image_name)
